% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeropuertos}
\alias{aeropuertos}
\title{Aeropuertos de La Pampa}
\format{
\subsection{\code{aeropuertos}}{

Un conjunto de datos geográfico con 2 filas y 5 variables:
\describe{
\item{ID}{Identificador del aeropuerto}
\item{Nombre}{Nombre del aeropuerto}
\item{Telefono}{Número de teléfono de contacto}
\item{Direccion}{Dirección del aeropuerto}
\item{geometry}{Geometría de tipo punto con la ubicación del aeropuerto}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
aeropuertos_la_pampa
}
\description{
Datos geográficos de aeropuertos en la Provincia de La Pampa.
}
\keyword{datasets}
