% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuencas_hidricas}
\alias{cuencas_hidricas}
\title{Cuencas hídricas de La Pampa}
\format{
\subsection{\code{cuencas_hidricas}}{

Un conjunto de datos geográfico con 6 filas y 6 variables:
\describe{
\item{FID_cuenca}{Identificador de la cuenca}
\item{NOMBRE}{Nombre de la cuenca hídrica}
\item{PROVINCIA}{Provincia}
\item{SISTEMA}{Sistema hídrico al que pertenece}
\item{SUPERFICIE}{Superficie de la cuenca}
\item{geometry}{Geometría de tipo polígono con el área de la cuenca}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
cuencas_hidricas_la_pampa
}
\description{
Datos geográficos de las cuencas hídricas de la Provincia de La Pampa.
}
\keyword{datasets}
