% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cursos_de_agua}
\alias{cursos_de_agua}
\title{Cursos de agua de La Pampa}
\format{
\subsection{\code{cursos_de_agua}}{

Un conjunto de datos geográfico con 1099 filas y 5 variables:
\describe{
\item{TIPO}{Tipo de curso de agua}
\item{NOMBRE}{Nombre del curso de agua}
\item{REGIMEN}{Régimen hídrico}
\item{Longitud}{Longitud del curso}
\item{geometry}{Geometría de tipo línea con el trazado del curso de agua}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
cursos_de_agua_la_pampa
}
\description{
Datos geográficos de los cursos de agua en la Provincia de La Pampa.
}
\keyword{datasets}
