% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departamentos}
\alias{departamentos}
\title{Departamentos de La Pampa}
\format{
\subsection{\code{departamentos}}{

Un conjunto de datos con 22 filas y 3 variables:
\describe{
\item{idDepartamento}{Identificador unico del departamento}
\item{nombreDepartamento}{Nombre del departamento}
\item{geomTDepartamento}{poligonos}
}
}
}
\usage{
departamentos
}
\description{
Divisiones politicas departamentales de La Pampa en formato geografico de poligonos.
}
\keyword{datasets}
