% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejidos_municipales_la_pampa}
\alias{ejidos_municipales_la_pampa}
\title{Ejidos municipales de La Pampa}
\format{
\subsection{\code{ejidos_municipales_la_pampa}}{

Un conjunto de datos geográfico con 79 filas y 46 variables:
\describe{
\item{Microregio}{Microrregión}
\item{Ejido}{Nombre del ejido municipal}
\item{Num_Ejido}{Número del ejido}
\item{Superficie}{Superficie del ejido}
\item{VERD_INV}{Verdeos de invierno (ha)}
\item{PASTURAS}{Pasturas (ha)}
\item{MONTE}{Monte (ha)}
\item{CAMPONAT}{Campo natural (ha)}
\item{AGRICULT}{Agricultura (ha)}
\item{RASTROJO}{Rastrojos (ha)}
\item{TRIGO_HA}{Trigo - hectáreas}
\item{TRIGO_QQ}{Trigo - quintales}
\item{MAIZ_HA}{Maíz - hectáreas}
\item{MAIZ_QQ}{Maíz - quintales}
\item{GIRAS_HA}{Girasol - hectáreas}
\item{GIRAS_QQ}{Girasol - quintales}
\item{SOJA_HA}{Soja - hectáreas}
\item{SOJA_QQ}{Soja - quintales}
\item{BOVINO}{Total bovinos}
\item{VACAS}{Vacas}
\item{VAQUILLO}{Vaquillonas}
\item{TERNEROS}{Terneros}
\item{TERNERAS}{Terneras}
\item{NOVILLIT}{Novillitos}
\item{NOVILLOS}{Novillos}
\item{TORITOS}{Toritos}
\item{TOROS}{Toros}
\item{OVINOS}{Total ovinos}
\item{OVEJAS}{Ovejas}
\item{CORDEROS}{Corderos}
\item{CORDERAS}{Corderas}
\item{BORREGOS}{Borregos}
\item{BORREGAS}{Borregas}
\item{CAPONES}{Capones}
\item{CARNEROS}{Carneros}
\item{PORCINOS}{Total porcinos}
\item{CERDAS}{Cerdas}
\item{LECHONES}{Lechones}
\item{CACHORRO}{Cachorros}
\item{CAPONESP}{Capones porcinos}
\item{PADRILLO}{Padrillos}
\item{CAPRINOS}{Total caprinos}
\item{MADRES}{Madres}
\item{CHIVITOS}{Chivitos}
\item{CASTRONE}{Castrones}
\item{Ind_Especi}{Índice de especialización}
\item{geometry}{Geometría de tipo polígono con el área del ejido}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
ejidos_municipales_la_pampa
}
\description{
Datos geográficos de los ejidos municipales de la Provincia de La Pampa con información
agropecuaria y ganadera.
}
\keyword{datasets}
