% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{espejos_de_agua}
\alias{espejos_de_agua}
\title{Espejos de agua de La Pampa}
\format{
\subsection{\code{espejos_de_agua}}{

Un conjunto de datos geográfico con 1086 filas y 5 variables:
\describe{
\item{TIPO}{Tipo de espejo de agua}
\item{NOMBRE}{Nombre del espejo de agua}
\item{REGIMEN}{Régimen hídrico}
\item{Superficie}{Superficie del espejo de agua}
\item{geometry}{Geometría de tipo polígono con el área del espejo de agua}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
espejos_de_agua
}
\description{
Datos geográficos de los espejos de agua en la Provincia de La Pampa.
}
\keyword{datasets}
