% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evp_Penman_Monteith}
\alias{evp_Penman_Monteith}
\title{Evapotranspiración de referencia (Penman-Monteith) de La Pampa}
\format{
\subsection{\code{evp_Penman_Monteith}}{

Un conjunto de datos geográfico con 3 filas y 2 variables:
\describe{
\item{EVP_REF_AN}{Evapotranspiración de referencia anual (mm)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
evp_Penman_Monteith
}
\description{
Datos geográficos de evapotranspiración de referencia anual calculada mediante
el método de Penman-Monteith para la Provincia de La Pampa.
}
\keyword{datasets}
