% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evp_real_thrnthwaite_la_pampa}
\alias{evp_real_thrnthwaite_la_pampa}
\title{Evapotranspiración real (Thornthwaite) de La Pampa}
\format{
\subsection{\code{evp_real_thrnthwaite_la_pampa}}{

Un conjunto de datos geográfico con 4 filas y 2 variables:
\describe{
\item{EVP_Real_A}{Evapotranspiración real anual (mm)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
evp_real_thrnthwaite_la_pampa
}
\description{
Datos geográficos de evapotranspiración real anual calculada mediante
el método de Thornthwaite para la Provincia de La Pampa.
}
\keyword{datasets}
