% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferrocarriles}
\alias{ferrocarriles}
\title{Ferrocarriles de La Pampa}
\format{
\subsection{\code{ferrocarriles}}{

Un conjunto de datos geográfico con 41 filas y 6 variables:
\describe{
\item{NOMBRE}{Nombre de la línea ferroviaria}
\item{TROCHA}{Tipo de trocha}
\item{HOJA}{Hoja cartográfica}
\item{Estado}{Estado de la línea}
\item{Tipo}{Tipo de línea}
\item{geometry}{Geometría de tipo línea con el trazado del ferrocarril}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
ferrocarriles_la_pampa
}
\description{
Datos geográficos de las líneas ferroviarias en la Provincia de La Pampa.
}
\keyword{datasets}
