% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fisonomia}
\alias{fisonomia}
\title{Informacion sobre la fisonomia relevada en los censos.}
\format{
\subsection{\code{fisonomia}}{

Conjunto de datos con 338 filas y 4 columnas:
\describe{
\item{idInventario}{Identificador único}
\item{fisonomia}{aspecto y rasgos (fisonómicos) y físicos visualmente evidenciables de las comunidades vegetales}
\item{alturaCoberturaFisonomia}{Cerrado, Muy Denso, Denso, Abierto, Muy Abierto, Ralo}
\item{Observacion}{Notas sobre el censo}
}
}
}
\source{
\url{https://recursosnaturales.lapampa.edu.ar/index2.html}
}
\usage{
fisonomia
}
\description{
Los datos corresponden a más de 300 censos de vegetación
realizados durante 1975 a 1979, utilizados para la caracterización
de los recursos naturales de la Provincia de La Pampa.
}
\keyword{datasets}
