% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fisonomia_2008_la_pampa_v2}
\alias{fisonomia_2008_la_pampa_v2}
\title{Fisonomía de la vegetación 2008 de La Pampa}
\format{
\subsection{\code{fisonomia_2008_la_pampa_v2}}{

Un conjunto de datos geográfico con 1276 filas y 4 variables:
\describe{
\item{LCCS_COD3}{Código LCCS de clasificación de cobertura del suelo}
\item{categoria}{Categoría de fisonomía}
\item{Superficie}{Superficie del área}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
fisonomia_2008_la_pampa_v2
}
\description{
Datos geográficos de la fisonomía de la vegetación en la Provincia de La Pampa según
clasificación LCCS (Land Cover Classification System) del año 2008.
}
\keyword{datasets}
