% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formacion}
\alias{formacion}
\title{Informacion sobre la formacion relevada en los censos.}
\format{
\subsection{\code{formacion}}{

Conjunto de datos con 907 filas y 3 columnas:
\describe{
\item{idInventario}{Identificador único}
\item{formacion}{clasificación de la vegetación según sea graminosa o leñosa}
\item{alturaCoberturaFormacion}{Cerrado, Muy Denso, Denso, Abierto, Muy Abierto, Ralo}
}
}
}
\source{
\url{https://recursosnaturales.lapampa.edu.ar/index2.html}
}
\usage{
formacion
}
\description{
Los datos corresponden a más de 300 censos de vegetación
realizados durante 1975 a 1979, utilizados para la caracterización
de los recursos naturales de la Provincia de La Pampa.
}
\keyword{datasets}
