% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horas_frio_anuales}
\alias{horas_frio_anuales}
\title{Horas de frío anuales de La Pampa}
\format{
\subsection{\code{horas_frio_anuales}}{

Un conjunto de datos geográfico con 4 filas y 3 variables:
\describe{
\item{Horas}{Número de horas de frío anuales}
\item{ID}{Identificador}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
horas_frio_anuales_la_pampa
}
\description{
Datos geográficos de las horas de frío anuales (temperatura por debajo de 7°C)
en la Provincia de La Pampa.
}
\keyword{datasets}
