% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventario}
\alias{inventario}
\title{Datos primarios del inventario de recursos naturales de La Pampa.}
\format{
\subsection{\code{inventario}}{

A data frame with 29,386 rows and 26 columns:
\describe{
\item{idInventario}{Identificador único}
\item{censo}{Numero de la planilla en papel correspondiente al censo}
\item{comunidad}{nivel de organización natural que incluye todas las poblaciones de especies diferentes de un área dada y en un tiempo dado, las interacciones entre ellas y con el ambiente}
\item{lugarInventario}{descripcion del lugar donde se realizo el censo}
\item{relieveInventario}{conjunto de irregularidades que presenta la superficie terrestre}
\item{sueloInventario}{}
\item{fechaInventario}{fecha en que se realizo el censo}
\item{idDepartamento}{identificador unico del departamento de La Pampa}
\item{tipoComunidad}{}
\item{observador}{profesionales que realizaron el censo a campo}
\item{formacion}{clasificación de la vegetación según sea graminosa o leñosa}
\item{alturaCoberturaFormacion}{Cerrado, Muy Denso, Denso, Abierto, Muy Abierto, Ralo}
\item{fisonomia}{aspecto y rasgos (fisonómicos) y físicos visualmente evidenciables de las comunidades vegetales}
\item{alturaCoberturaFisonomia}{Cerrado, Muy Denso, Denso, Abierto, Muy Abierto, Ralo}
\item{Observacion}{Notas sobre el censo}
\item{descripcionCobertura}{valore +, -1, 1, 1 - 2, 2, 2 - 3, 3, 3 - 4, 4}
\item{porcentajeCobertura}{valores 2.50, 5.00, 7.50, 12.50, 17.50, 25.00, 37.50, 50.00, 62.50}
\item{estrato}{cada uno de los niveles en que pueden ordenarse las especies de una comunidad en el plano vertical.}
\item{nombreCientificoEspecie}{nombre cientifico de la especie relevada en el censo}
\item{tipoBiologico}{tipo biologico de la especia, valores C, Cactacea, Enredadera, F, G, H, P}
\item{especie}{unidad básica de la clasificación biológica. Conjunto de organismos (poblaciones) naturales capaces de entrecruzarse y de producir descendencia fértil.}
\item{preferencia}{preferencia de la especie, valores Intermedia, No preferida, Preferida}
\item{estacionalidad}{estacionalidad de la especie, valores Invernal, Primavero-estival}
\item{grupoFuncional}{grupo funcional de la especie, valores Herbácea, Leñosa}
\item{ciclo}{ciclo de la especie, valores Anual, Perenne}
\item{geometry}{lacalizacion donde se realizo el censo}
}
}
}
\source{
\url{https://recursosnaturales.lapampa.edu.ar/index2.html}
}
\usage{
inventario
}
\description{
Los datos corresponden a más de 300 censos de vegetación
y 150 calicatas de suelos realizados durante 1975 a 1979,
utilizados para la caracterización de los recursos naturales
de la Provincia de La Pampa.
}
\keyword{datasets}
