% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventario_flora}
\alias{inventario_flora}
\title{Inventario de la flora presente en los censos}
\format{
\subsection{\code{inventario_flora}}{

Conjunto de datos con 7699 filas y 11 columnas:
\describe{
\item{idInventario}{Identificador único}
\item{descripcionCobertura}{valore +, -1, 1, 1 - 2, 2, 2 - 3, 3, 3 - 4, 4}
\item{porcentajeCobertura}{valores 2.50, 5.00, 7.50, 12.50, 17.50, 25.00, 37.50, 50.00, 62.50}
\item{estrato}{cada uno de los niveles en que pueden ordenarse las especies de una comunidad en el plano vertical.}
\item{nombreCientificoEspecie}{nombre cientifico de la especie relevada en el censo}
\item{especie}{unidad básica de la clasificación biológica. Conjunto de organismos (poblaciones) naturales capaces de entrecruzarse y de producir descendencia fértil.}
\item{tipoBiologico}{tipo biologico de la especia, valores C, Cactacea, Enredadera, F, G, H, P}
\item{preferencia}{preferencia de la especie, valores Intermedia, No preferida, Preferida}
\item{estacionalidad}{estacionalidad de la especie, valores Invernal, Primavero-estival}
\item{grupoFuncional}{grupo funcional de la especie, valores Herbácea, Leñosa}
\item{ciclo}{ciclo de la especie, valores Anual, Perenne}
}
}
}
\source{
\url{https://recursosnaturales.lapampa.edu.ar/index2.html}
}
\usage{
inventario_flora
}
\description{
Los datos corresponden a más de 300 censos de vegetación
realizados durante 1975 a 1979, utilizados para la caracterización
de los recursos naturales de la Provincia de La Pampa.
}
\keyword{datasets}
