% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventario_vegetal}
\alias{inventario_vegetal}
\title{Informacion sobre los censos}
\format{
\subsection{\code{inventario_vegetal}}{

Conjunto de datos con 351 filas y 10 columnas:
\describe{
\item{idInventario}{Identificador único}
\item{censo}{Numero de la planilla en papel correspondiente al censo}
\item{comunidad}{nivel de organización natural que incluye todas las poblaciones de especies diferentes de un área dada y en un tiempo dado, las interacciones entre ellas y con el ambiente}
\item{lugarInventario}{descripcion del lugar donde se realizo el censo}
\item{relieveInventario}{conjunto de irregularidades que presenta la superficie terrestre}
\item{sueloInventario}{}
\item{fechaInventario}{fecha en que se realizo el censo}
\item{idDepartamento}{identificador unico del departamento de La Pampa}
\item{tipoComunidad}{}
\item{observador}{profesionales que realizaron el censo a campo}
}
}
}
\source{
\url{https://recursosnaturales.lapampa.edu.ar/index2.html}
}
\usage{
inventario_vegetal
}
\description{
Los datos corresponden a más de 300 censos de vegetación
realizados durante 1975 a 1979, utilizados para la caracterización
de los recursos naturales de la Provincia de La Pampa.
}
\keyword{datasets}
