% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localizacion_censos}
\alias{localizacion_censos}
\title{Localizacion geografica de los censos}
\format{
\subsection{\code{localizacion_censos}}{

Un conjunto de datos con 423 filas y 2 variables:
\describe{
\item{censo}{Identificado unico del censo}
\item{geometry}{localizacion de los censos en formato de puntos}
}
}
}
\usage{
localizacion_censos
}
\description{
Latitud y longitud con la ubicacion geografica de los censos de campo.
}
\keyword{datasets}
