% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear_censos.R
\name{mapear_censos}
\alias{mapear_censos}
\title{Mapear ubicacion de censos con ubicacion geografica}
\usage{
mapear_censos()
}
\value{
un objeto ggplot2 con un mapa que muestra la ubicacion geografica de
todos los censos realizados en La Pampa.
}
\description{
Mapear ubicacion de censos con ubicacion geografica
}
