% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precipitaciones_anuales}
\alias{precipitaciones_anuales}
\title{Precipitaciones anuales de La Pampa}
\format{
\subsection{\code{precipitaciones_anuales}}{

Un conjunto de datos geográfico con 5 filas y 3 variables:
\describe{
\item{ID}{Identificador}
\item{MM}{Precipitación anual en milímetros}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
precipitaciones_anuales_la_pampa
}
\description{
Datos geográficos de las precipitaciones anuales promedio en la Provincia de La Pampa.
}
\keyword{datasets}
