% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primera_helada_1_5m}
\alias{primera_helada_1_5m}
\title{Primera helada a 1.5m de altura en La Pampa}
\format{
\subsection{\code{primera_helada_1_5m}}{

Un conjunto de datos geográfico con 7 filas y 2 variables:
\describe{
\item{FECHA}{Fecha promedio de la primera helada (formato día/mes)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
primera_helada_1_5m
}
\description{
Datos geográficos con la fecha promedio de la primera helada del año medida a
1.5 metros de altura en la Provincia de La Pampa.
}
\keyword{datasets}
