% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regiones_geomorfologicas_la_pampa_v2}
\alias{regiones_geomorfologicas_la_pampa_v2}
\title{Regiones geomorfológicas de La Pampa}
\format{
\subsection{\code{regiones_geomorfologicas_la_pampa_v2}}{

Un conjunto de datos geográfico con 18 filas y 9 variables:
\describe{
\item{ID}{Identificador de la región}
\item{SUBREGION}{Nombre de la subregión geomorfológica}
\item{Superficie}{Superficie de la región}
\item{ALTIMETRIA}{Características altimétricas}
\item{RELIEVE}{Tipo de relieve}
\item{LITOLOGIA}{Características litológicas}
\item{DRENAJE}{Características del drenaje}
\item{SECTORIZ}{Sectorización}
\item{geometry}{Geometría de tipo polígono con el área de la región}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
regiones_geomorfologicas_la_pampa_v2
}
\description{
Datos geográficos de las regiones geomorfológicas de la Provincia de La Pampa.
}
\keyword{datasets}
