% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regiones_hidricas}
\alias{regiones_hidricas}
\title{Regiones hídricas de La Pampa}
\format{
\subsection{\code{regiones_hidricas}}{

Un conjunto de datos geográfico con 13 filas y 4 variables:
\describe{
\item{NOMBRE}{Nombre de la región hídrica}
\item{CODIGO}{Código de la región}
\item{Superficie}{Superficie de la región}
\item{geometry}{Geometría de tipo polígono con el área de la región}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
regiones_hidricas
}
\description{
Datos geográficos de las regiones hídricas de la Provincia de La Pampa.
}
\keyword{datasets}
