% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_especie.R
\name{reporte_especie}
\alias{reporte_especie}
\title{Lista todas las especies presentes en un censo}
\usage{
reporte_especie(numero_censo)
}
\arguments{
\item{numero_censo}{}
}
\value{
tabla
}
\description{
Lista todas las especies presentes en un censo con el calculo
del porcentaje por especie en cada estrato presente. Si la
especie no esta presente en un estrato se presenta NA.
}
\details{
Si un estrato no tiene presencia de ninguna especie no se
muestra en la tabla.
}
\examples{
reporte_especie(44)

}
