% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_estacion.R
\name{reporte_estacion}
\alias{reporte_estacion}
\title{Listado de cantidad de especies por estacion}
\usage{
reporte_estacion()
}
\value{
tabla
}
\description{
Listado del total de especies presentes en cada censo
realizado agrupadas por la estación del año en que
crece cada especie.
}
\examples{
reporte_estacion()

}
