% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suelos}
\alias{suelos}
\title{Suelos de La Pampa}
\format{
\subsection{\code{suelos}}{

Un conjunto de datos geográfico con 261 filas y 30 variables:
\describe{
\item{ID_CARTO}{Identificador cartográfico}
\item{Cod_RRNN}{Código de recursos naturales}
\item{Cod_Atlas}{Código del atlas}
\item{Superficie}{Superficie del área}
\item{PROVINCIA}{Provincia}
\item{SIMB_TAXO}{Símbolo taxonómico}
\item{SIMB_ORD}{Símbolo del orden}
\item{TIPO_UC}{Tipo de unidad cartográfica}
\item{PAISAJE}{Tipo de paisaje}
\item{REGION_IP}{Región del índice de productividad}
\item{ZONA_IP}{Zona del índice de productividad}
\item{IP_CARTO}{Índice de productividad cartográfico}
\item{USDA_CAR}{Clasificación USDA cartográfica}
\item{SUB_USDA}{Sub-clasificación USDA}
\item{LIM1_CAR}{Limitación 1 cartográfica}
\item{LIM2_CAR}{Limitación 2 cartográfica}
\item{LIM3_CAR}{Limitación 3 cartográfica}
\item{NRO_COMP}{Número de componente}
\item{PAISAJE2}{Paisaje secundario}
\item{TAXON_1}{Taxón 1}
\item{FAMILIA_1}{Familia 1}
\item{FASE_1}{Fase 1}
\item{TAXON_2}{Taxón 2}
\item{FAMILIA_2}{Familia 2}
\item{FASE_2}{Fase 2}
\item{TAXON_3}{Taxón 3}
\item{FAMILIA_3}{Familia 3}
\item{FASE_3}{Fase 3}
\item{Orden}{Orden del suelo}
\item{geometry}{Geometría de tipo polígono con el área}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
suelos_la_pampa
}
\description{
Datos geográficos de los suelos de la Provincia de La Pampa con clasificación taxonómica.
}
\keyword{datasets}
