% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatura_media}
\alias{temperatura_media}
\title{Temperatura media anual de La Pampa}
\format{
\subsection{\code{temperatura_media}}{

Un conjunto de datos geográfico con 4 filas y 3 variables:
\describe{
\item{Id}{Identificador}
\item{TEMP}{Temperatura media anual (°C)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
temperatura_media
}
\description{
Datos geográficos de la temperatura media anual en la Provincia de La Pampa.
}
\keyword{datasets}
