% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatura_mes_mas_calido_la_pampa}
\alias{temperatura_mes_mas_calido_la_pampa}
\title{Temperatura del mes más cálido de La Pampa}
\format{
\subsection{\code{temperatura_mes_mas_calido_la_pampa}}{

Un conjunto de datos geográfico con 8 filas y 3 variables:
\describe{
\item{Id}{Identificador}
\item{TEMP}{Temperatura media del mes más cálido (°C)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
temperatura_mes_mas_calido_la_pampa
}
\description{
Datos geográficos de la temperatura media del mes más cálido (enero)
en la Provincia de La Pampa.
}
\keyword{datasets}
