% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatura_mes_mas_frio}
\alias{temperatura_mes_mas_frio}
\title{Temperatura del mes más frío de La Pampa}
\format{
\subsection{\code{temperatura_mes_mas_frio}}{

Un conjunto de datos geográfico con 2 filas y 2 variables:
\describe{
\item{TEMP}{Temperatura media del mes más frío (°C)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
temperatura_mes_mas_frio_la_pampa
}
\description{
Datos geográficos de la temperatura media del mes más frío (julio)
en la Provincia de La Pampa.
}
\keyword{datasets}
