% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ultima_helada_0_05m}
\alias{ultima_helada_0_05m}
\title{Última helada a 0.05m de altura en La Pampa}
\format{
\subsection{\code{ultima_helada_0_05m}}{

Un conjunto de datos geográfico con 10 filas y 2 variables:
\describe{
\item{FECHA}{Fecha promedio de la última helada (formato día/mes)}
\item{geometry}{Geometría de tipo polígono}
}
}
}
\source{
Casagrande, Vergara y Bellini. "Cartas agroclimáticas actuales de temperaturas,
heladas y lluvia de la provincia de La Pampa (Argentina)".
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
ultima_helada_0_05m_la_pampa
}
\description{
Datos geográficos con la fecha promedio de la última helada del año medida a
0.05 metros de altura en la Provincia de La Pampa.
}
\keyword{datasets}
