% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegetacion_rn}
\alias{vegetacion_rn}
\title{Vegetación de recursos naturales de La Pampa}
\format{
\subsection{\code{vegetacion_rn}}{

Un conjunto de datos geográfico con 336 filas y 5 variables:
\describe{
\item{VEGETACION}{Tipo de vegetación}
\item{Superficie}{Superficie del área}
\item{FISONOMIAS}{Fisonomías presentes}
\item{CARACTERIS}{Características de la vegetación}
\item{geometry}{Geometría de tipo polígono con el área}
}
}
}
\source{
Roberto, Zinda; Martínez Uncal, María Celeste. Bases para el Ordenamiento Territorial.
Publicación Técnica N° 87. ISSN 0325-2132. Mayo 2012. Ediciones INTA.
Descargado de GeoINTA con licencia CC-BY.
}
\usage{
vegetacion_rn
}
\description{
Datos geográficos de la vegetación de la Provincia de La Pampa.
}
\keyword{datasets}
