% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departamentos}
\alias{departamentos}
\title{Departamentos de La Pampa}
\format{
:
\describe{
\item{idDepartamento}{Identificador unico del departamento}
\item{nombreDepartamento}{Nombre del departamento}
\item{geomTDepartamento}{polignos}
}
}
\usage{
departamentos
}
\description{
Divisiones politicas departamentales de La Pampa en formato geografico de poligonos.
}
\keyword{datasets}
