% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear_especie.R
\name{mapear_especie}
\alias{mapear_especie}
\title{Mapear ubicacion de censos con una especie}
\usage{
mapear_especie(nombre_cientifico)
}
\arguments{
\item{nombre_cientifico}{Nombre cientifico de la especie a buscar (caracter).
Debe coincidir exactamente con los valores en la columna
nombreCientificoEspecie de \code{inventario_flora}.}
}
\value{
Un objeto ggplot2 con un mapa que muestra la ubicacion geografica de
los censos donde se encontro la especie.
Si la especie no se encuentra en ningun censo, retorna un mapa sin puntos
}
\description{
Filtra los censos que contienen una especie determinada y retorna
los datos geograficos de ubicacion de esos censos.
}
